Reverted base_build change for Boost for Android  (#56)
* Removed suppress logs for Boost build

I can't tell how this will actually affect Android builds until this is in master. For now, since Android builds are optional, this is OK, but in the future, we probably want a stop gap of some sort to see how this would affect things for MADARA builds. For instance, we should probably have Android builds in GAMS that replicate the MADARA builds. Otherwise, we might merge things and break MADARA builds and we won't know until after the merge and rebuild on MADARA side.